+++ src/main/java/com/iredko/wowcraft/intr/GenericManagerInterface.java
@@ -8,11 +8,11 @@ public interface GenericManagerInterface<T> {
 
     public T update(T object);
 
-    public T findById(Long id);
+    public T findById(Integer id);
 
     public void delete(T object);
 
     public void insert(T object);
 
-    public boolean exist(Long id);
+    public boolean exist(Integer id);
 }
